LPS-80124 (bookmarks-uad) Use BaseModelUADDisplay for UADDisplay base implementation
