+++ modules/apps/collaboration/bookmarks/bookmarks-uad/src/main/java/com/liferay/bookmarks/uad/display/BaseBookmarksEntryUADDisplay.java
@@ -1,0 +1,57 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.bookmarks.uad.display;
+
+import com.liferay.bookmarks.model.BookmarksEntry;
+import com.liferay.bookmarks.uad.constants.BookmarksUADConstants;
+import com.liferay.user.associated.data.display.BaseModelUADDisplay;
+
+import java.util.Locale;
+
+/**
+ * Provides the base implementation for the bookmarks entry UAD display.
+ *
+ * <p>
+ * This implementation exists only as a container for the default methods
+ * generated by ServiceBuilder. All custom methods should be put in
+ * {@link BookmarksEntryUADDisplay}.
+ * </p>
+ *
+ * @author William Newbury
+ */
+public abstract class BaseBookmarksEntryUADDisplay
+	extends BaseModelUADDisplay<BookmarksEntry> {
+
+	public String getApplicationName() {
+		return BookmarksUADConstants.APPLICATION_NAME;
+	}
+
+	public String[] getDisplayFieldNames() {
+		return new String[] {"name", "description", "url"};
+	}
+
+	public String getKey() {
+		return BookmarksUADConstants.CLASS_NAME_BOOKMARKS_ENTRY;
+	}
+
+	public String getTypeDescription() {
+		return "A link to another page or website";
+	}
+
+	public String getTypeName(Locale locale) {
+		return "BookmarksEntry";
+	}
+
+}
